Changed 'C' code
Another efficient means of identifying duplicates in an integer array,is
to implement a HashTable (HashSet). We use Linked List chaining
toresolve collisions.HashTable size is not dynamically set here.
Set_Size = 5000 seems toyield the least execution time for the 18 test
cases on leetcode.com for this problem.
18 / 18 test cases passed
Runtime: 8 ms